+++ internal/stockrequest/types.go
@@ -1,6 +1,6 @@
 package stockrequest
 
 type DispatchItem struct {
-	VariantID string `json:"variant_id"`
-	Qty       int    `json:"dispatch_qty"`
+	VariantID string  `json:"variant_id"`
+	Qty       float64 `json:"dispatch_qty"`
 }
